Refactor AzNavRail to a single, cohesive DSL-style API
This commit refactors the AzNavRail library to resolve the architectural conflict between two competing APIs. The legacy UI implementation has been removed, and its functionality has been merged into the main `AzNavRail` composable.

The key changes are:
- The `AzNavRailUi.kt` file and the `ui` directory have been deleted.
- The logic for handling expanded/collapsed states, header, footer, and different item types has been moved into private composables within `AzNavRail.kt`.
- The public API (`AzNavRail` and `AzNavRailScope`) remains consistent, with some additional optional parameters for more control.
- The `README.md` has been updated to reflect the new signature of the `AzNavRail` composable.

This results in a single, cohesive, and maintainable library with a unified DSL-style API.
